# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tether API",
    "description": "\n# tether API\n\ntether helps you hold yourself accountable to keep your phone away from your bedroom at night.\n\n## Overview\n\nThis API runs on a Raspberry Pi and provides:\n\n1. **Proximity Detection**: Check if your phone is near the Raspberry Pi via Bluetooth\n2. **Emergency Passes**: A limited number of monthly passes for legitimate exceptions\n3. **Configuration**: Manage Bluetooth devices and settings\n\n## For AI Agents (MCP)\n\nIf you're accessing this API via MCP tools:\n\n- **checkProximity**: Verify the phone is in its designated spot. Returns `is_nearby: true` when close.\n- **getPasses**: Check how many emergency passes remain this month.\n- **usePass**: Use when user has legitimate reason (on-call, emergency). Requires a reason.\n- **getPassHistory**: Review past pass usage to identify patterns.\n\n## Design Philosophy\n\n- **Lazy evaluation**: Bluetooth checks only happen when requested\n- **Intentional friction**: Passes require reasons to encourage mindfulness\n- **Delayed effects**: Pass count changes only apply next month to prevent gaming\n",
    "contact": {
      "name": "Jeffrey",
      "email": "jeffrey@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Local tether server"
    }
  ],
  "paths": {
...
    "/config/bluetooth": {
      "put": {
        "tags": [
          "config"
        ],
        "summary": "Update Bluetooth target device",
        "description": "Updates the Bluetooth device to track for proximity detection.",
        "operationId": "updateBluetooth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBluetoothRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Bluetooth configuration updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateBluetoothResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid Bluetooth address format"
          }
        }
      }
    },
    "/config/onboarding/complete": {
      "put": {
        "tags": [
          "config"
        ],
        "summary": "Complete onboarding",
        "description": "Marks the initial onboarding as complete. Requires Bluetooth target to be configured first.",
        "operationId": "completeOnboarding",
        "responses": {
          "200": {
            "description": "Onboarding completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteOnboardingResponse"
                }
              }
            }
          },
          "400": {
            "description": "Already completed"
          },
          "424": {
            "description": "Prerequisites not met"
          }
        }
      }
    },
    "/config/passes": {
      "put": {
        "tags": [
          "config"
        ],
        "summary": "Update passes per month",
        "description": "Updates the number of emergency passes allowed per month. If passes have already been used this month, the change will take effect next month.",
        "operationId": "updatePassesPerMonth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePassesPerMonthRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Passes per month updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatePassesPerMonthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid value"
          }
        }
      }
    },
    "/config/timezone": {
      "put": {
        "tags": [
          "config"
        ],
        "summary": "Update timezone",
        "description": "Updates the timezone used for pass reset calculations.",
        "operationId": "updateTimezone",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTimezoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Timezone updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateTimezoneResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid timezone"
          }
        }
      }
    },
    "/config/wifi": {
      "put": {
        "tags": [
          "config"
        ],
        "summary": "Update WiFi networks",
        "description": "Updates the list of WiFi networks. Each network needs an SSID, password, and whether it's the primary network. Exactly one network should be marked as primary.",
        "operationId": "updateWifi",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWifiRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "WiFi configuration updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateWifiResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid WiFi configuration"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BluetoothConfigResponse": {
        "type": "object",
        "description": "Bluetooth configuration in response.",
        "required": [
          "target_address",
          "target_name",
          "rssi_threshold",
          "is_configured"
        ],
        "properties": {
          "is_configured": {
            "type": "boolean",
            "description": "Whether a real device has been configured (not placeholder).",
            "example": true
          },
          "rssi_threshold": {
            "type": "integer",
            "format": "int32",
            "description": "RSSI threshold for proximity detection.",
            "example": -60
          },
          "target_address": {
            "type": "string",
            "description": "Bluetooth MAC address of target device.",
            "example": "AA:BB:CC:DD:EE:FF"
          },
          "target_name": {
            "type": "string",
            "description": "User-friendly name of the device.",
            "example": "iPhone 15 Pro"
          }
        },
        "example": {
          "is_configured": true,
          "rssi_threshold": -60,
          "target_address": "AA:BB:CC:DD:EE:FF",
          "target_name": "iPhone 15 Pro"
        }
      },
      "CompleteOnboardingResponse": {
        "type": "object",
        "description": "Response after completing onboarding.",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Message about onboarding status."
          },
          "success": {
            "type": "boolean",
            "description": "Whether onboarding was completed."
          }
        }
      },
...
      "UpdateBluetoothRequest": {
        "type": "object",
        "description": "Request to update Bluetooth target device.",
        "required": [
          "target_address",
          "target_name"
        ],
        "properties": {
          "rssi_threshold": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Optional RSSI threshold (-100 to 0 dBm). Defaults to -60.",
            "example": -60
          },
          "target_address": {
            "type": "string",
            "description": "Bluetooth MAC address (XX:XX:XX:XX:XX:XX format).",
            "example": "AA:BB:CC:DD:EE:FF"
          },
          "target_name": {
            "type": "string",
            "description": "User-friendly name for the device.",
            "example": "iPhone 15 Pro"
          }
        },
        "example": {
          "rssi_threshold": -60,
          "target_address": "AA:BB:CC:DD:EE:FF",
          "target_name": "iPhone 15 Pro"
        }
      },
      "UpdateBluetoothResponse": {
        "type": "object",
        "description": "Response after updating Bluetooth configuration.",
        "required": [
          "success",
          "bluetooth"
        ],
        "properties": {
          "bluetooth": {
            "$ref": "#/components/schemas/BluetoothConfigResponse",
            "description": "Updated configuration."
          },
          "success": {
            "type": "boolean",
            "description": "Whether the update was successful."
          }
        }
      },
      "UpdatePassesPerMonthRequest": {
        "type": "object",
        "description": "Request to update passes per month.",
        "required": [
          "per_month"
        ],
        "properties": {
          "per_month": {
            "type": "integer",
            "format": "int32",
            "description": "Number of passes per month (0-31).",
            "example": 3,
            "maximum": 31,
            "minimum": 0
          }
        },
        "example": {
          "per_month": 3
        }
      },
      "UpdatePassesPerMonthResponse": {
        "type": "object",
        "description": "Response after updating passes per month.",
        "required": [
          "success",
          "per_month",
          "pending",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "Message explaining when the change takes effect."
          },
          "pending": {
            "type": "boolean",
            "description": "Whether the change is pending (will apply next month).",
            "example": true
          },
          "per_month": {
            "type": "integer",
            "format": "int32",
            "description": "Updated passes per month value.",
            "example": 3,
            "minimum": 0
          },
          "success": {
            "type": "boolean",
            "description": "Whether the update was successful."
          }
        }
      },
      "UpdateTimezoneRequest": {
        "type": "object",
        "description": "Request to update timezone.",
        "required": [
          "timezone"
        ],
        "properties": {
          "timezone": {
            "type": "string",
            "description": "IANA timezone name (e.g., \"America/Los_Angeles\").",
            "example": "America/Los_Angeles"
          }
        },
        "example": {
          "timezone": "America/Los_Angeles"
        }
      },
      "UpdateTimezoneResponse": {
        "type": "object",
        "description": "Response after updating timezone.",
        "required": [
          "success",
          "timezone"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the update was successful."
          },
          "timezone": {
            "type": "string",
            "description": "Updated timezone.",
            "example": "America/Los_Angeles"
          }
        }
      },
      "UpdateWifiRequest": {
        "type": "object",
        "description": "Request to update WiFi networks.",
        "required": [
          "networks"
        ],
        "properties": {
          "networks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WifiNetworkConfig"
            },
            "description": "List of WiFi networks to configure."
          }
        },
        "example": {
          "networks": [
            {
              "is_primary": true,
              "password": "secret123",
              "ssid": "HomeNetwork"
            },
            {
              "is_primary": false,
              "password": "backup456",
              "ssid": "BackupNetwork"
            }
          ]
        }
      },
      "UpdateWifiResponse": {
        "type": "object",
        "description": "Response after updating WiFi configuration.",
        "required": [
          "success",
          "networks_count"
        ],
        "properties": {
          "networks_count": {
            "type": "integer",
            "description": "Number of networks configured.",
            "minimum": 0
          },
          "primary_ssid": {
            "type": [
              "string",
              "null"
            ],
            "description": "The SSID of the primary network."
          },
          "success": {
            "type": "boolean",
            "description": "Whether the update was successful."
          }
        }
      },
...
      "WifiNetworkConfig": {
        "type": "object",
        "description": "A WiFi network configuration.",
        "required": [
          "ssid",
          "password",
          "is_primary"
        ],
        "properties": {
          "is_primary": {
            "type": "boolean",
            "description": "Whether this is the primary network.",
            "example": true
          },
          "password": {
            "type": "string",
            "description": "Network password.",
            "example": "supersecret123"
          },
          "ssid": {
            "type": "string",
            "description": "Network SSID (name).",
            "example": "HomeNetwork"
          }
        },
        "example": {
          "is_primary": true,
          "password": "supersecret123",
          "ssid": "HomeNetwork"
        }
      }
    }
  },
...
}